Use supplied loader when cloning worlds (#51)
* Use clone loader when cloning worlds

Cloning a world from file to db would save the cloned world
to db and then on shutdown it would also save the in-memory
world to file since that was the loader of the world which was
cloned. Instead set the loader to the one supplied to the clone method

* Use clone loader when cloning worlds

Cloning a world from file to db would save the cloned world
to db and then on shutdown it would also save the in-memory
world to file since that was the loader of the world which was
cloned. Instead set the loader to the one supplied to the clone method